modified implementation for the 50% brightness default
The old implementation capped the Unicorn HAT brightness to 128 (of 255) and this couldn't be overridden by users of the unicornhat module. The brightness() default value of b=0.2 had no effect until someone actually called the function. And 0.2*128 is 25.6, well below the low brightness warning treshold of 30 ...
Why not allow a range of 255 (which is fine if only a few pixels are lit), but set an effective default by calling brightness(0.5) in the module?
Problem: it changes the behaviour of applications that already call brightness() assuming a result of b*128.