Added two examples for unicornHAT
I propose two python 3 scripts for the examples folder.

First, is candle_unicornhat.py, which adapts the candle example from unicorn-hat-hd to the unicornHAT, by resizing the animation from 16x16 to 8x8 pixels.
Based on: https://github.com/pimoroni/unicorn-hat-hd/blob/master/examples/candle.py

Second, candle_with_timer.py extends the above example by adding a timer with a fade-out effect. The purpose is make it easy to use the candle animation as a scheduled go-to-sleep light, with soothing warm colors and pleasant connotations of hearth and home. The fade-out effect is intended to simulate tapering the candle down, reducing the size and brightness of the flame gradually before the script ends and the flame shuts off. The timer length and fade-out length in minutes are individual variables, so each are easy to adjust within the script, and extensively commented.